broke Main into multiple classes
